+++ src/bakery_manager_mobile/lib/pages/routes.dart
@@ -6,8 +6,11 @@ import 'package:bakery_manager_mobile/pages/sales_page.dart';
 import 'package:bakery_manager_mobile/pages/task_page.dart';
 import 'package:bakery_manager_mobile/pages/login.dart';
 import 'package:bakery_manager_mobile/pages/account_registration.dart';
+import 'package:bakery_manager_mobile/pages/my_account.dart';
 import 'package:flutter/material.dart';
 import 'package:bakery_manager_mobile/assets/constants.dart';
+import 'package:bakery_manager_mobile/pages/add_recipe_page.dart';
+import 'package:bakery_manager_mobile/pages/recipe_details_page.dart';
 
 Map<String, WidgetBuilder> appRoutes = {
   homePageRoute: (context) => const MyHomePage(),
@@ -18,4 +21,7 @@ Map<String, WidgetBuilder> appRoutes = {
   salesPageRoute: (context) => const SalesPage(),
   loginPageRoute: (context) => const LoginPage(),
   registrationPageRoute: (context) => const CreateAccountPage(),
+  addRecipePageRoute: (context) => const AddRecipePage(),
+  recipeDetailsPageRoute: (context) => const RecipeDetailPage(),
+  myAccountRoute: (context) => const MyAccountPage(),
 };